çölAyak
We have formed a project team for individuals with gluten sensitivity and created a website with the aim of raising awareness about this condition. We prepared a database from approved gluten-free product websites. In the initial phase, we developed a website that showcases products from stores selling gluten-free products within a city.